% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_MonoClust.PULS.R
\name{as_MonoClust.PULS}
\alias{as_MonoClust.PULS}
\title{Coerce a PULS Object to MonoClust Object}
\usage{
\method{as_MonoClust}{PULS}(x, ...)
}
\arguments{
\item{x}{A PULS object to be coerced to MonoClust object.}

\item{...}{For extensibility.}
}
\value{
A MonoClust object coerced from PULS object.
}
\description{
An implementation of the \code{\link[monoClust:as_MonoClust]{monoClust::as_MonoClust()}} S3 method for PULS
object. The purpose of this is to reuse plotting and printing functions from
\link{monoClust} package.
}
\seealso{
\link[monoClust:MonoClust.object]{monoClust::MonoClust.object} and \link{PULS.object}
}
