% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puls.R
\name{checkem}
\alias{checkem}
\title{First Gate Function}
\usage{
checkem(
  toclust.fd,
  frame,
  cloc,
  dist,
  dsubs,
  dsubsname,
  weights,
  minbucket,
  minsplit,
  spliton,
  method
)
}
\arguments{
\item{toclust.fd}{A functional data object (i.e., having class \code{fd}) created
from \code{fda} package. See \code{\link[fda:fd]{fda::fd()}}.}

\item{frame}{The split tree transferred as data frame.}

\item{cloc}{Vector of current cluster membership.}

\item{dist}{Distance matrix of all observations in the data.}

\item{dsubs}{Distance matrix calculated on each subregion. A
three-dimensional matrix.}

\item{dsubsname}{Subregion names.}

\item{weights}{(Currently unused) Weights on observations.}

\item{minbucket}{The minimum number of data points in one cluster allowed.}

\item{minsplit}{The minimum number of observations that must exist in a node
in order for a split to be attempted.}

\item{spliton}{Restrict the partitioning on a specific set of subregions.}

\item{method}{The clustering method you want to run in each subregion. Can be
chosen between \code{pam} and \code{ward}.}
}
\value{
It is not supposed to return anything because global environment was
used. However, if there is nothing left to split, it returns 0 to tell the
caller to stop running the loop.
}
\description{
This function checks what are available nodes to split and then call
\code{find_split()} on each node, then decide which node creates best split, and
call \code{splitter()} to perform the split.
}
\keyword{internal}
