% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PULS.R
\name{create_labels}
\alias{create_labels}
\title{Create Labels for Split Variables}
\usage{
create_labels(x, digits, ...)
}
\arguments{
\item{x}{PULS result object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Optional arguments to \code{\link[=abbreviate]{abbreviate()}},}
}
\value{
A list containing two elements:
\itemize{
\item \code{names}: A named vector of labels corresponding to range's names
(at vector names).
\item \code{labels}: Vector of labels of splitting rules to be displayed.
}
}
\description{
This function prints variable's labels for a \code{PULS} tree.
}
\keyword{internal}
