% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdistmatrix.R
\name{fdistmatrix}
\alias{fdistmatrix}
\title{Distance between Functional Objects}
\usage{
fdistmatrix(yfd, subrange, distmethod)
}
\arguments{
\item{yfd}{A functional data object as the result of \code{fda} package.}

\item{subrange}{A vector of two values indicating the value range of
functional object to calculate on.}

\item{distmethod}{The method for calculating the distance matrix. Choose
between \code{usc} and \code{manual}.}
}
\value{
A distance matrix with diagonal value and the upper half.
}
\description{
Calculate the distance between functional objects over the defined range.
}
\examples{
library(fda)
# Examples taken from fda::Data2fd()
data(gait)
# Function only works on two dimensional data
gait <- gait[, , 1]
gaitbasis3 <- create.fourier.basis(nbasis = 5)
gaitfd3 <- Data2fd(gait, basisobj = gaitbasis3)

fdistmatrix(gaitfd3, c(0.2, 0.4), "usc")
}
