% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wave.R
\name{plot.wave}
\alias{plot.wave}
\title{Plot the Partitioned Functional Wave by PULS}
\usage{
\method{plot}{wave}(fd, puls, member = T, medoids = T, separate = F, ...)
}
\arguments{
\item{fd}{the fdSmooth object, created by fda::smooth.basis() function}

\item{puls}{the output of PULS algorithm}

\item{member}{whether or not making cluster's members different colors}

\item{medoids}{whether or not highlighting the medoid of each cluster}

\item{separate}{if \code{TRUE}, the function will create separate plot for each cluster}

\item{...}{other optional plotting arguments}
}
\description{
After partitioning using PULS, this function can plot the functional waves and
color different clusters as well as their medoids
}
