% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dat.R
\name{read_dat}
\alias{read_dat}
\title{Read a dat file}
\usage{
read_dat(file)
}
\arguments{
\item{file}{the file path to the input file}
}
\value{
a data.frame with data extracted from the input dat file
}
\description{
Reads a dat file and creates a data frame from it.
}
\examples{
dat_filepath = system.file("dat/CR3000_Rain.dat", package = "puls")
my_dat = read_dat(dat_filepath)
head(my_dat)
}
