% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ghg.R
\name{read_ghg}
\alias{read_ghg}
\title{Read a ghg file}
\usage{
read_ghg(file)
}
\arguments{
\item{file}{the file path to the input file}
}
\value{
a data.frame with data extracted from the input ghg file
}
\description{
Reads a ghg file and creates a data frame from it.
}
\examples{
ghg_filepath = system.file("ghg/2020-02-19T220000_AIU-0874.ghg", package = "puls")
my_ghg = read_ghg(ghg_filepath)
head(my_ghg)
}
