% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoothed_arctic}
\alias{smoothed_arctic}
\title{Discrete Form of Smoothed Functional Form of Arctic Data}
\format{
A data frame with 39 rows corresponding to 39 years (1979 to 1986,
1989 to 2019) and 366 columns:
\describe{
\item{year}{Year of data.}
\item{day_1 to day_365}{365 days in a year.}
}
}
\usage{
smoothed_arctic
}
\description{
Raw Arctic data were smoothed and then transformed into functional data using
\code{fda} package. To overcome the difficulty of exporting an \code{fda} object in a
package, the object was discretized into a data set with 365 columns
corresponding to 365 days a year and 39 rows corresponding to
39 years. The years are from 1979 to 1986, then from 1989 to 2018. The years
1978, 1987, and 1988 were removed because the measurements were not complete.
}
\seealso{
NOAA's raw data at \link{arctic_2019} and the code to generate this data
in data-raw/ folder of source code.
}
\keyword{datasets}
