% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soccer_intervals}
\alias{soccer_intervals}
\title{NOAA's Arctic Sea Daily Ice Extend Data}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{start}{years of available data (1978--2019)}
\item{end}{month (01--12)}
}
}
\source{
\url{https://nsidc.org/data/G02135/versions/3}
}
\usage{
soccer_intervals
}
\description{
A data set containing the daily ice extent at Arctic Sea from 1978 to 2019,
collected by National Oceanic and Atmospheric Administration (NOAA)
}
\keyword{datasets}
