% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{validate_nan}
\alias{validate_nan}
\alias{validate_}
\alias{validate_min}
\alias{validate_max}
\alias{validate_longterm}
\title{Title}
\usage{
validate_nan(y)

validate_min(y, param_df)

validate_max(y, param_df)

validate_longterm(y, param_df, howmanysd = 3)
}
\arguments{
\item{y}{a}

\item{param_df}{b}

\item{howmanysd}{c}
}
\value{
d
}
\description{
Title
}
