% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{cleaner}
\alias{cleaner}
\title{Prepares Input Data}
\usage{
cleaner(x, site_code = NULL)
}
\arguments{
\item{Path}{to a \code{.dat} file}
}
\value{
A reformatted data frame with a TIMESTAMP, RECORD, name, value, parameter, and site
}
\description{
This function is used iternally in the \code{validator()} function.
}
\examples{
dat_filepath = system.file("dat/CR3000_Rain.dat", package = "puls")
cleaner(dat_filepath, "BR")
}
