% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PULS.R
\name{plot.PULS}
\alias{plot.PULS}
\title{Plot PULS Splitting Rule Tree}
\usage{
\method{plot}{PULS}(x, margin, which, text = TRUE, cols = NULL, ...)
}
\arguments{
\item{x}{PULS object.}

\item{margin}{An extra fraction of white space to leave around the borders of
the tree. (Long labels sometimes get cut off by the default computation).}

\item{which}{Labeling modes, which are:
\itemize{
\item 1: only splitting variable names are shown, no splitting rules.
\item 2: only splitting rules to the left branches are shown.
\item 3: only splitting rules to the right branches are shown.
\item 4 (default): splitting rules are shown on both sides of branches.
}}

\item{text}{Whether to print the labels on the tree.}

\item{cols}{Whether to use color or not.}

\item{...}{Arguments to be passed to \code{\link[graphics:plot.default]{graphics::plot.default()}} and
\code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
A plot of splitting order.
}
\description{
Print the PULS tree in the form of dendrogram.
}
\examples{
library(fda)
}
