% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PULS.R
\name{plot.PULS}
\alias{plot.PULS}
\title{Title}
\usage{
\method{plot}{PULS}(x, margin, which, abbrev = 4, text = TRUE,
  cols = NULL, ...)
}
\arguments{
\item{cols}{Whether to use color or not}

\item{...}{}
}
\description{
Title
}
