% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PULS.R
\name{print.PULS}
\alias{print.PULS}
\title{Title}
\usage{
\method{print}{PULS}(x, abbrev = 0, spaces = 2, digits = options("digits")$digits, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
