% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puls-package.R
\docType{package}
\name{puls-package}
\alias{puls}
\alias{puls-package}
\title{puls: Misc Functions of the Laboratory of Bioclimatology, PULS Poznan}
\description{
A set of tools to read and validate data obtained by the Laboratory of Bioclimatology, PULS Poznan.
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
