% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puls-package.R
\docType{package}
\name{puls-package}
\alias{puls}
\alias{puls-package}
\title{puls: Partitioning Using Local Subregions}
\description{
A method of clustering functional data using
    subregion information of the curves. It is intended to supplement the
    'fda' and 'fda.usc' packages in functional data object clustering. It
    also facilitates the printing and plotting of the results in a tree
    format and limits the partitioning candidates into a specific set of
    subregions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vinhtantran.github.io/puls/}
  \item \url{https://github.com/vinhtantran/puls}
  \item Report bugs at \url{https://github.com/vinhtantran/puls/issues}
}

}
\author{
\strong{Maintainer}: Tan Tran \email{vinhtantran@gmail.com} (\href{https://orcid.org/0000-0001-9881-6339}{ORCID})

Authors:
\itemize{
  \item Mark Greenwood \email{greenwood@montana.edu} (\href{https://orcid.org/0000-0001-6933-1201}{ORCID})
}

}
\keyword{internal}
