% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PULS-package.R
\docType{package}
\name{PULS-package}
\alias{PULS}
\alias{PULS-package}
\title{PULS: Partitioning Using Local Subregions}
\description{
Partitioning using local subregions (PULS) is a
    clustering technique designed to explore subregions of functional data
    for information to split the curves into clusters.
}
\author{
\strong{Maintainer}: Tan Tran \email{vinhtantran@gmail.com} (\href{https://orcid.org/0000-0001-9881-6339}{ORCID})

Authors:
\itemize{
  \item Mark Greenwood \email{greenwood@montana.edu} (\href{https://orcid.org/0000-0001-6933-1201}{ORCID})
}

}
\keyword{internal}
