% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PULS.R
\name{PULS}
\alias{PULS}
\title{Partitioning Using Local Subregions (PULS)}
\usage{
PULS(toclust.fd, method = c("pam", "ward"), intervals = c(0, 1),
  spliton = NULL, distmethod = c("usc", "manual"),
  labels = toclust.fd$fdnames[2]$reps,
  nclusters = length(toclust.fd$fdnames[2]$reps), minbucket = 2,
  minsplit = 4)
}
\arguments{
\item{toclust.fd}{data set, whereas rows are functional entities and columns are
variables}

\item{method}{the clustering method you want to run in each subregion. Can be
chosen between \code{pam} and \code{ward}}

\item{intervals}{a data set (or matrix) with rows are intervals and columns
are the beginning and ending indices of of the interval}

\item{spliton}{restrict the partitioning on a specific set of subregions}

\item{distmethod}{the method for calculating the distance matrix}

\item{labels}{the name of entities}

\item{nclusters}{the number of clusters}

\item{minbucket}{the minimum number of data points in one cluster allowed}

\item{minsplit}{the minimum size of a cluster that can still be considered to
be a split candidate}
}
\description{
PULS function for functional data (only used when you know that the data
shouldn't be convereted into functional because it's already smooth, e.g.
your data are step function)
}
