% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PULS.R
\name{PULS}
\alias{PULS}
\title{Partitioning Using Local Subregions (PULS)}
\usage{
PULS(
  toclust.fd,
  method = c("pam", "ward"),
  intervals = c(0, 1),
  spliton = NULL,
  distmethod = c("usc", "manual"),
  labels = toclust.fd$fdnames[2]$reps,
  nclusters = length(toclust.fd$fdnames[2]$reps),
  minbucket = 2,
  minsplit = 4
)
}
\arguments{
\item{toclust.fd}{A functional data object as the result of \code{fda} package.}

\item{method}{The clustering method you want to run in each subregion. Can be
chosen between \code{pam} and \code{ward}.}

\item{intervals}{A data set (or matrix) with rows are intervals and columns
are the beginning and ending indexes of of the interval.}

\item{spliton}{Restrict the partitioning on a specific set of subregions.}

\item{distmethod}{The method for calculating the distance matrix. Choose
between \code{"usc"} and \code{"manual"}.}

\item{labels}{The name of entities.}

\item{nclusters}{The number of clusters.}

\item{minbucket}{The minimum number of data points in one cluster allowed.}

\item{minsplit}{The minimum size of a cluster that can still be considered to
be a split candidate.}
}
\description{
PULS function for functional data (only used when you know that the data
shouldn't be converted into functional because it's already smooth, e.g.
your data are step function)
}
\examples{
library(fda)
NORDER <- 1
splinebasis <- create.bspline.basis(rangeval = c(1, 275),
                                    norder = NORDER,
                                    breaks = 1:275)

# No need for any more smoothing
fdParobj <- fdPar(fdobj = splinebasis,
                  Lfdobj = 0,
                  lambda = .000001)

yfd <- smooth.basis(argvals = 1:ncol(soccer.wide[, -1]),
                    y = t(as.matrix(soccer.wide[, -1])),
                    fdParobj = fdParobj)
puls_obj <- PULS(yfd$fd,
                 intervals = intervals,
                 labels = soccer.wide[, 1],
                 nclusters = 4)
plot(puls_obj)
}
\seealso{
\code{\link[fda:is.fd]{fda::is.fd()}}
}
