% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{validate_nan}
\alias{validate_nan}
\alias{validate_}
\alias{validate_min}
\alias{validate_max}
\alias{validate_longterm}
\title{Validate Several Separate Border Cases}
\usage{
validate_nan(y)

validate_min(y, param_df)

validate_max(y, param_df)

validate_longterm(y, param_df, howmanysd = 3)
}
\arguments{
\item{y}{A list with two data frames - \code{correct_df} and \code{error_df}}

\item{param_df}{A data frame with basic parameters' ranges}

\item{howmanysd}{Specifies how many standard deviations from the provided mean are accepted}
}
\value{
A list with two data frames - \code{correct_df} and \code{error_df}.
\code{correct_df} stores data without any detected problems and
\code{error_df} stores data with detected problems together with
a specified problem description.
}
\description{
This function is used iternally in the \code{validator()} function.
}
