% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validator}
\alias{validator}
\title{Title}
\usage{
validator(
  x,
  parameter,
  start_date = as.Date("2019-07-25"),
  end_date = as.Date("2019-07-30")
)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
Title
}
